@@ -6,14 +6,20 @@ authors = ["91fabbai91 <fabian.baier@gmx.de>"]
 readme = "README.md"
 
 [tool.poetry.dependencies]
-python = "^3.8"
-httplib2 = "^0.20.4"
-lxml = "^4.9.1"
-PyYAML = "^6.0"
-pytest = "^7.1.3"
-pytest-mock = "^3.8.2"
-bidict = "^0.22.0"
+python = "^3.10.12"
+httplib2 = "^0.22.0"
+lxml = "^4.9.3"
+PyYAML = "^6.0.1"
+pydantic = "^2.2.1"
+python-dotenv = "^1.0.0"
+pydantic-settings = "^2.0.3"
+pre-commit = "^3.3.3"
+pylint = "^2.17.5"
 
+
+[tool.poetry.group.test.dependencies]
+pytest = "^7.4.0"
+pytest-mock = "^3.11.1"
 
 [build-system]
 requires = ["poetry-core"]
