@@ -1,12 +1,16 @@
 import logging
 import re
+import unicodedata
+
 import http_connection
-import parsing_utils
+from lxml import html
+from offer import Offer
 
 
-class Marketplace(object):
+class Marketplace:
     FEE_PERCENTAGE = 0.1
     MINIMAL_DISCOUNT = 0.01
+
     def __init__(self, http_connection: http_connection.HTTPConnection) -> None:
         self.__logger = logging.getLogger(self.__class__.__name__)
         self.__http_connection = http_connection
@@ -14,80 +18,98 @@ class Marketplace(object):
 
     def get_all_tradeable_products(self):
         """
-        Returns the IDs of all tradable products.
+        Returns the IDs of all tradeable products.
         """
         self.update_all_tradeable_products()
         return self.__tradeable_product_ids
-    
+
     def update_all_tradeable_products(self):
         content = self.__http_connection.get_all_tradeable_products().decode()
-        tradeableProducts = re.findall(r'markt\.php\?order=p&v=([0-9]{1,3})&filter=1', content)
-        for i in range(0, len(tradeableProducts)):
-            tradeableProducts[i] = int(tradeableProducts[i])
-        self.__tradeable_product_ids = tradeableProducts       
-        return tradeableProducts
+        tradeable_products = re.findall(
+            r"markt\.php\?order=p&v=([0-9]{1,3})&filter=1", content
+        )
 
-    def get_all_offers_of_product(self, id):
+        tradeable_products = [int(x) for x in tradeable_products]
+        self.__tradeable_product_ids = tradeable_products
+        return tradeable_products
+
+    def get_all_offers_of_product(self, identifier: int) -> list:
         """
         Determines all offers of a product.
         """
         self.update_all_tradeable_products()
-        
-        if self.__tradeable_product_ids != None \
-           and \
-           id in self.__tradeable_product_ids:
 
-            list_offers = self.__http_connection.get_offers_from_product(id)
-        
-        else: #Product is not tradeable
-            list_offers = None
-            
+        if (
+            self.__tradeable_product_ids is not None
+            and identifier in self.__tradeable_product_ids
+        ):
+            list_offers = self.__get_offers_from_product(identifier)
+
+        else:  # Product is not tradeable
+            list_offers = []
+
         return list_offers
 
-    def get_cheapest_offer(self, id):
+    def __get_offers_from_product(self, identifier: int) -> list:
+        """
+        Gibt eine Liste mit allen Angeboten eines Produkts zurück.
+        """
+        list_offers = []
+
+        next_page = True
+        page_number = 1
+        while next_page:
+            next_page = False
+            content = self.__http_connection.execute_market_command(
+                identifier=identifier, page=page_number
+            )
+            html_tree = html.fromstring(content)
+            table = html_tree.xpath("//body/div/table")[0]
+            if table[1][0].text == "Keine Angebote":
+                list_offers = []
+            else:
+                # range von 1 bis länge-1, da erste Zeile Überschriften sind
+                # und die letzte Weiter/Zurück. Falls es mehrere seiten gibt.
+                list_offers = get_amounts_and_prices_from_table(table)
+
+                for element in table[::-1][0]:
+                    if "weiter" in element.text:
+                        next_page = True
+                        page_number = page_number + 1
+
+        return list_offers
+
+    def get_cheapest_offer(self, identifier: int, filtered_username: str = "") -> float:
         """
         Determines the most favorable offer of a product.
         """
 
-        list_offers = self.get_all_offers_of_product(id)
-        
-        if len(list_offers) >= 1 \
-           and \
-           list_offers != None:
-            
-            return list_offers[0][1]
-        
-        else: #No Offers
-            return None
+        list_offers = self.get_all_offers_of_product(identifier)
+        if list_offers is not None:
+            list_offers = list(
+                filter(lambda x: x.seller == filtered_username, list_offers)
+            )
+            if len(list_offers) > 0:
+                return list_offers[0].price
+        return float("inf")
 
-    def sell_on_market(self, item_id: int , price: float, number: int):
+    def sell_on_market(self, item_id: int, price: float, number: int):
         self.__http_connection.sell_on_market(item_id, price, number)
 
-    def find_big_gap_in_product_offers(self, id, npcPrice):
-        """
-        Identifies a large gap (> 10%) between offers and returns it.
-        """
-        
-        list_offers = self.get_all_offers_of_product(id)
-        list_prices = []
 
-        if (list_offers != None):
-            
-            #Collect all prices in one list
-            for element in list_offers:
-                list_prices.append(element[1])
-            
-            if (npcPrice != None and id != 0): #id != 0: Do not sort coins
-                iList = reversed(range(0, len(list_prices)))
-                for i in iList:
-                    if list_prices[i] > npcPrice:
-                        del list_prices[i]
-            
-            gaps = []
-            #At least two entries are required for comparison.
-            if (len(list_prices) >= 2):
-                for i in range(0, len(list_prices)-1):
-                    if (((list_prices[i+1] / 1.1) - list_prices[i]) > 0.0):
-                        gaps.append([list_prices[i], list_prices[i+1]])
-            
-            return gaps
+def get_amounts_and_prices_from_table(table):
+    list_offers = []
+    for i in range(1, len(table) - 1):
+        amount = table[i][0].text
+        amount = amount.replace(".", "")
+
+        price = table[i][3].text
+        price = unicodedata.normalize("NFKD", price)
+        price, _ = price.split()
+        price = price.replace(".", "")
+        price = price.replace(",", ".")
+        product = table[i][1][0].text
+        seller = table[i][2][0].text
+        offer = Offer(product=product, amount=amount, price=price, seller=seller)
+        list_offers.append(offer)
+    return list_offers
